use float for amount fields in swap

+++ broker/model/swap.go
@@ -7,13 +7,13 @@ type Swap struct {
 	PoolID                    uint32  `json:"pool_id"`
 	SwapRequester             string  `json:"swap_requester"`
 	OfferCoinDenom            string  `json:"offer_coin_denom"`
-	OfferCoinAmount           int64   `json:"offer_coin_amount"`
+	OfferCoinAmount           float64 `json:"offer_coin_amount"`
 	DemandCoinDenom           string  `json:"demand_coin_denom"`
-	ExchangedDemandCoinAmount int64   `json:"exchanged_demand_coin_amount"`
-	TransactedCoinAmount      int64   `json:"transacted_coin_amount"`
-	RemainingOfferCoinAmount  int64   `json:"remaining_offer_coin_amount"`
-	OfferCoinFeeAmount        int64   `json:"offer_coin_fee_amount"`
-	OrderExpiryHeight         int64   `json:"order_expiry_height"`
+	ExchangedDemandCoinAmount float64 `json:"exchanged_demand_coin_amount"`
+	TransactedCoinAmount      float64 `json:"transacted_coin_amount"`
+	RemainingOfferCoinAmount  float64 `json:"remaining_offer_coin_amount"`
+	OfferCoinFeeAmount        float64 `json:"offer_coin_fee_amount"`
+	OrderExpiryHeight         float64 `json:"order_expiry_height"`
 	ExchangedCoinFeeAmount    float64 `json:"exchanged_coin_fee_amount"`
 	OrderPrice                float64 `json:"order_price"`
 	SwapPrice                 float64 `json:"swap_price"`
